move parseAsPlayer() to PlayerUtils
