+++ src/main/java/factionsplusplus/utils/StringUtils.java
@@ -5,9 +5,6 @@ import java.util.UUID;
 
 import org.bukkit.ChatColor;
 
-import org.bukkit.Bukkit;
-import org.bukkit.OfflinePlayer;
-
 public class StringUtils {
     public static Integer parseAsInteger(String string) {
         try {
@@ -30,19 +27,6 @@ public class StringUtils {
         final String[] falseValues = {"no", "off", "false", "n"};
         if (Arrays.asList(trueValues).contains(string.toLowerCase())) return true;
         if (Arrays.asList(falseValues).contains(string.toLowerCase())) return false;
-        return null;
-    }
-
-    @SuppressWarnings("deprecation")
-    public static OfflinePlayer parseAsPlayer(String playerName) {
-        final OfflinePlayer player = Bukkit.getOfflinePlayer(playerName);
-        if (player.hasPlayedBefore() || Bukkit.getPlayer(player.getUniqueId()) != null) return player;
-        return null;
-    }
-
-    public static OfflinePlayer parseAsPlayer(UUID playerUUID) {
-        final OfflinePlayer player = Bukkit.getOfflinePlayer(playerUUID);
-        if (player.hasPlayedBefore() || Bukkit.getPlayer(player.getUniqueId()) != null) return player;
         return null;
     }
